% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_duration}
\alias{check_duration}
\title{Check duration}
\usage{
check_duration(duration)
}
\arguments{
\item{duration}{(ms)}
}
\value{
TRUE if invalid
}
\description{
Check duration
}
\examples{
check_duration(501)
}
