% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_velocity}
\alias{check_velocity}
\title{Check velocity}
\usage{
check_velocity(velocity)
}
\arguments{
\item{velocity}{(m/s)}
}
\value{
TRUE if invalid
}
\description{
Check velocity
}
\examples{
check_velocity(5.1)
}
