% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{convert_to_acceleration}
\alias{convert_to_acceleration}
\title{Convert from duration to acceleration}
\usage{
convert_to_acceleration(delta_velocity, time)
}
\arguments{
\item{delta_velocity}{m/s}

\item{time}{ms}
}
\value{
cm/s2
}
\description{
Convert from duration to acceleration
}
\examples{
converto_to_acceleration(2.1, 70)
}
