% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{convert_to_meter}
\alias{convert_to_meter}
\title{Convert to meter from cm}
\usage{
convert_to_meter(value, convert)
}
\arguments{
\item{value}{cm}

\item{convert}{logical if conversion needed}
}
\value{
value in m
}
\description{
Convert to meter from cm
}
