% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{convert_to_time}
\alias{convert_to_time}
\title{Convert from acceleration to duration}
\usage{
convert_to_time(delta_velocity, acceleration)
}
\arguments{
\item{delta_velocity}{m/s}

\item{acceleration}{m/s2}
}
\value{
Duration in ms
}
\description{
Convert from acceleration to duration
}
\examples{
convert_to_time(2.1, 30)
}
