% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewave_velocity_fx_time.R
\name{ewave_velocity_fx_time}
\alias{ewave_velocity_fx_time}
\title{Velocity of motion as function of time for E-waves}
\usage{
ewave_velocity_fx_time(.t, C, K, x0)
}
\arguments{
\item{.t}{time as numeric}

\item{C}{Viscoelasticity (g/s)}

\item{K}{Stiffness (g/s2)}

\item{x0}{Load (m)}
}
\value{
vt, velocity as a function of time
}
\description{
In the PDF framework the following expressions
describe the velocity of motion as function of time
}
\examples{

}
