% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewave_velocity_fx_time.R
\name{ewave_velocity_fx_time_data}
\alias{ewave_velocity_fx_time_data}
\title{Vectorized over time: Velocity of motion as function of time for E-waves}
\usage{
ewave_velocity_fx_time_data(C, K, x0)
}
\arguments{
\item{C}{Viscoelasticity (g/s)}

\item{K}{Stiffness (g/s2)}

\item{x0}{Load (m)}
}
\description{
Vectorized over time: Velocity of motion as function of time for E-waves
}
