% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{export_data}
\alias{export_data}
\title{Export data helper}
\usage{
export_data(data, summary, selected_rows, choice, excl_col)
}
\arguments{
\item{data}{\code{data.frame} of raw data}

\item{summary}{\code{data.frame} of summary data}

\item{selected_rows}{\code{vector}}

\item{choice}{\code{character} What to export c("selection", "all", "summary")}

\item{excl_col}{charcter \code{vector} of columns to exclude from export}
}
\value{
\code{data.frame}
}
\description{
Export data helper
}
