% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fzero.R
\name{fzero}
\alias{fzero}
\title{fzero}
\usage{
fzero(f, x, ..., maxiter = 100, tol = .Machine$double.eps^(1/2))
}
\arguments{
\item{f}{Function to solve}

\item{x}{Initial value, specified as a numeric or a 2-element vector}

\item{...}{Further options for solution process}

\item{maxiter}{Maximum number of iterations (default: 100)}

\item{tol}{Tolerance settings (default: .Machine$double.eps^(1/2))}
}
\description{
Root of nonlinear function ported from matlab
}
