% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_initial_pdf_parameters.R
\name{generate_c_k_x0}
\alias{generate_c_k_x0}
\title{Generate viscoelastic energyloss, stiffness, and load}
\usage{
generate_c_k_x0(AT, DT, Epeak)
}
\arguments{
\item{AT}{E acceleration time (milliseconds)}

\item{DT}{E decelleration time (milliseconds)}

\item{Epeak}{E peak velocity (meter/second)}
}
\value{
Stiffness, viscoelsticity, Load
}
\description{
Generate viscoelastic energyloss, stiffness, and load
}
\examples{
generate_c_k_x0(AT = 50, DT = 200, Epeak = 1.1)
}
