% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pdf_parameters.R
\name{generate_pdf_parameters}
\alias{generate_pdf_parameters}
\title{Generate Parameterized Diastolic Filling parameters}
\usage{
generate_pdf_parameters(C, K, x0, Epeak, AT, DT)
}
\arguments{
\item{C}{Viscoelasticity (g/s)}

\item{K}{Stiffness (g/s2)}

\item{x0}{Load (cm)}

\item{Epeak}{E peak velocity (meter/second)}

\item{AT}{E acceleration time (milliseconds)}

\item{DT}{E decelleration time (milliseconds)}
}
\value{
Numerics
}
\description{
Generate Parameterized Diastolic Filling parameters
}
